songsensei/scripts/requirements.txt
httpx==0.25.0
asyncio
python-dotenv
loguru
